+++ cmd/export/main.go
@@ -1,24 +1,21 @@
 package main
 
 import (
-	"bytes"
+	"crypto/sha256"
 	"encoding/json"
 	"flag"
 	"fmt"
 	"github.com/MinterTeam/go-amino"
+	"github.com/MinterTeam/minter-go-node/cmd/export/types11"
 	"github.com/MinterTeam/minter-go-node/cmd/utils"
 	"github.com/MinterTeam/minter-go-node/core/state"
-	"github.com/pkg/errors"
-	"github.com/tendermint/tendermint/crypto"
-	"github.com/tendermint/tendermint/crypto/ed25519"
-	"github.com/tendermint/tendermint/crypto/secp256k1"
+	mtypes "github.com/MinterTeam/minter-go-node/core/types"
 	"github.com/tendermint/tendermint/libs/common"
 	"github.com/tendermint/tendermint/libs/db"
 	"github.com/tendermint/tendermint/types"
-	tmtime "github.com/tendermint/tendermint/types/time"
-	"golang.org/x/crypto/sha3"
-	"io/ioutil"
+	"io"
 	"log"
+	"os"
 	"time"
 )
 
@@ -28,14 +25,22 @@ var (
 	genesisTime = flag.Duration("genesis_time", 0, "genesis_time")
 )
 
+const (
+	genesisPath = "genesis.json"
+
+	maxSupply = "1000000000000000000000000000000000"
+)
+
 func main() {
 	defer func() {
 		if r := recover(); r != nil {
-			fmt.Println("Recovered in f", r)
+			fmt.Println("Error occurred", r)
 		}
 	}()
-	required := []string{"height", "chain_id", "genesis_time"}
+
 	flag.Parse()
+
+	required := []string{"height", "chain_id", "genesis_time"}
 	seen := make(map[string]bool)
 	flag.Visit(func(f *flag.Flag) { seen[f.Name] = true })
 	for _, req := range required {
@@ -59,34 +64,96 @@ func main() {
 		panic(err)
 	}
 
-	cdc := amino.NewCodec()
+	oldState := currentState.Export(*height)
+	newState := types11.AppState{
+		Note:         oldState.Note,
+		StartHeight:  oldState.StartHeight,
+		MaxGas:       oldState.MaxGas,
+		TotalSlashed: oldState.TotalSlashed.String(),
+	}
 
-	jsonBytes, err := cdc.MarshalJSONIndent(currentState.Export11(*height), "", "	")
+	for _, account := range oldState.Accounts {
+		newState.Accounts = append(newState.Accounts, mtypes.Account{
+			Address:      account.Address,
+			Balance:      account.Balance,
+			Nonce:        account.Nonce,
+			MultisigData: account.MultisigData,
+		})
+	}
+
+	for _, coin := range oldState.Coins {
+		newState.Coins = append(newState.Coins, types11.Coin{
+			Name:      coin.Name,
+			Symbol:    coin.Symbol,
+			Volume:    coin.Volume.String(),
+			Crr:       coin.Crr,
+			Reserve:   coin.ReserveBalance.String(),
+			MaxSupply: maxSupply,
+		})
+	}
+
+	for _, check := range oldState.UsedChecks {
+		newState.UsedChecks = append(newState.UsedChecks, check)
+	}
+
+	for _, ff := range oldState.FrozenFunds {
+		newState.FrozenFunds = append(newState.FrozenFunds, mtypes.FrozenFund{
+			Height:       ff.Height,
+			Address:      ff.Address,
+			CandidateKey: ff.CandidateKey,
+			Coin:         ff.Coin,
+			Value:        ff.Value,
+		})
+	}
+
+	for _, candidate := range oldState.Candidates {
+		newState.Candidates = append(newState.Candidates, mtypes.Candidate{
+			RewardAddress: candidate.RewardAddress,
+			OwnerAddress:  candidate.OwnerAddress,
+			TotalBipStake: candidate.TotalBipStake,
+			PubKey:        candidate.PubKey,
+			Commission:    candidate.Commission,
+			Stakes:        candidate.Stakes,
+			Status:        candidate.Status,
+		})
+	}
+
+	for _, validator := range oldState.Validators {
+		newState.Validators = append(newState.Validators, mtypes.Validator{
+			TotalBipStake: validator.TotalBipStake,
+			PubKey:        validator.PubKey,
+			AccumReward:   validator.AccumReward,
+			AbsentTimes:   validator.AbsentTimes,
+		})
+	}
+
+	cdc := amino.NewCodec()
+	jsonBytes, err := cdc.MarshalJSONIndent(newState, "", "	")
 	if err != nil {
 		panic(err)
 	}
 
-	appHash := [32]byte{}
+	appHash := currentState.Hash()
 
 	// Compose Genesis
-	genesis := GenesisDoc{
+	genesis := types11.GenesisDoc{
 		GenesisTime: time.Unix(0, 0).Add(*genesisTime),
 		ChainID:     *chainID,
-		ConsensusParams: &ConsensusParams{
-			Block: BlockParams{
+		ConsensusParams: &types11.ConsensusParams{
+			Block: types11.BlockParams{
 				MaxBytes:   10000000,
 				MaxGas:     100000,
 				TimeIotaMs: 1000,
 			},
-			Evidence: EvidenceParams{
+			Evidence: types11.EvidenceParams{
 				MaxAgeNumBlocks: 1000,
-				MaxAgeDuration:  24 * time.Hour, //todo
+				MaxAgeDuration:  24 * time.Hour,
 			},
-			Validator: ValidatorParams{
+			Validator: types11.ValidatorParams{
 				PubKeyTypes: []string{types.ABCIPubKeyTypeEd25519},
 			},
 		},
-		AppHash:  appHash[:],
+		AppHash:  appHash,
 		AppState: json.RawMessage(jsonBytes),
 	}
 
@@ -95,177 +162,24 @@ func main() {
 		panic(err)
 	}
 
-	if err := genesis.SaveAs("genesis.json"); err != nil {
+	if err := genesis.SaveAs(genesisPath); err != nil {
 		panic(err)
 	}
 
-	fmt.Println("OK")
-	fmt.Println(fmt.Sprintf("%x", sha3.Sum512([]byte(fmt.Sprintf("%v", genesis)))))
+	fmt.Printf("Ok\n%x\n", getSha256Hash(genesisPath))
 }
 
-type GenesisDoc struct {
-	GenesisTime     time.Time          `json:"genesis_time"`
-	ChainID         string             `json:"chain_id"`
-	ConsensusParams *ConsensusParams   `json:"consensus_params,omitempty"`
-	Validators      []GenesisValidator `json:"validators,omitempty"`
-	AppHash         common.HexBytes    `json:"app_hash"`
-	AppState        json.RawMessage    `json:"app_state,omitempty"`
-}
-type ConsensusParams struct {
-	Block     BlockParams     `json:"block"`
-	Evidence  EvidenceParams  `json:"evidence"`
-	Validator ValidatorParams `json:"validator"`
-}
-type BlockParams struct {
-	MaxBytes int64 `json:"max_bytes"`
-	MaxGas   int64 `json:"max_gas"`
-	// Minimum time increment between consecutive blocks (in milliseconds)
-	// Not exposed to the application.
-	TimeIotaMs int64 `json:"time_iota_ms"`
-}
-type GenesisValidator struct {
-	Address Address       `json:"address"`
-	PubKey  crypto.PubKey `json:"pub_key"`
-	Power   int64         `json:"power"`
-	Name    string        `json:"name"`
-}
-type Address = crypto.Address
-type EvidenceParams struct {
-	MaxAgeNumBlocks int64         `json:"max_age_num_blocks"` // only accept new evidence more recent than this
-	MaxAgeDuration  time.Duration `json:"max_age_duration"`
-}
-type ValidatorParams struct {
-	PubKeyTypes []string `json:"pub_key_types"`
-}
-
-const (
-	MaxChainIDLen = 50
-)
-
-func (genDoc *GenesisDoc) ValidateAndComplete() error {
-	if genDoc.ChainID == "" {
-		return errors.New("genesis doc must include non-empty chain_id")
-	}
-	if len(genDoc.ChainID) > MaxChainIDLen {
-		return errors.Errorf("chain_id in genesis doc is too long (max: %d)", MaxChainIDLen)
-	}
-
-	if genDoc.ConsensusParams == nil {
-		genDoc.ConsensusParams = DefaultConsensusParams()
-	} else if err := genDoc.ConsensusParams.Validate(); err != nil {
-		return err
-	}
-
-	for i, v := range genDoc.Validators {
-		if v.Power == 0 {
-			return errors.Errorf("the genesis file cannot contain validators with no voting power: %v", v)
-		}
-		if len(v.Address) > 0 && !bytes.Equal(v.PubKey.Address(), v.Address) {
-			return errors.Errorf("incorrect address for validator %v in the genesis file, should be %v", v, v.PubKey.Address())
-		}
-		if len(v.Address) == 0 {
-			genDoc.Validators[i].Address = v.PubKey.Address()
-		}
-	}
-
-	if genDoc.GenesisTime.IsZero() {
-		genDoc.GenesisTime = tmtime.Now()
-	}
-
-	return nil
-}
-func DefaultConsensusParams() *ConsensusParams {
-	return &ConsensusParams{
-		DefaultBlockParams(),
-		DefaultEvidenceParams(),
-		DefaultValidatorParams(),
-	}
-}
-func DefaultBlockParams() BlockParams {
-	return BlockParams{
-		MaxBytes:   22020096, // 21MB
-		MaxGas:     -1,
-		TimeIotaMs: 1000, // 1s
-	}
-}
-
-// DefaultEvidenceParams Params returns a default EvidenceParams.
-func DefaultEvidenceParams() EvidenceParams {
-	return EvidenceParams{
-		MaxAgeNumBlocks: 100000, // 27.8 hrs at 1block/s
-		MaxAgeDuration:  48 * time.Hour,
-	}
-}
-
-// DefaultValidatorParams returns a default ValidatorParams, which allows
-// only ed25519 pubkeys.
-func DefaultValidatorParams() ValidatorParams {
-	return ValidatorParams{[]string{types.ABCIPubKeyTypeEd25519}}
-}
-
-const (
-	ABCIPubKeyTypeEd25519   = "ed25519"
-	ABCIPubKeyTypeSr25519   = "sr25519"
-	ABCIPubKeyTypeSecp256k1 = "secp256k1"
-
-	MaxBlockSizeBytes = 104857600
-)
-
-var ABCIPubKeyTypesToAminoNames = map[string]string{
-	ABCIPubKeyTypeEd25519:   ed25519.PubKeyAminoName,
-	ABCIPubKeyTypeSr25519:   "tendermint/PubKeySr25519",
-	ABCIPubKeyTypeSecp256k1: secp256k1.PubKeyAminoName,
-}
-
-func (params *ConsensusParams) Validate() error {
-	if params.Block.MaxBytes <= 0 {
-		return errors.Errorf("block.MaxBytes must be greater than 0. Got %d",
-			params.Block.MaxBytes)
-	}
-	if params.Block.MaxBytes > MaxBlockSizeBytes {
-		return errors.Errorf("block.MaxBytes is too big. %d > %d",
-			params.Block.MaxBytes, MaxBlockSizeBytes)
-	}
-
-	if params.Block.MaxGas < -1 {
-		return errors.Errorf("block.MaxGas must be greater or equal to -1. Got %d",
-			params.Block.MaxGas)
-	}
-
-	if params.Block.TimeIotaMs <= 0 {
-		return errors.Errorf("block.TimeIotaMs must be greater than 0. Got %v",
-			params.Block.TimeIotaMs)
-	}
-
-	if params.Evidence.MaxAgeNumBlocks <= 0 {
-		return errors.Errorf("evidenceParams.MaxAgeNumBlocks must be greater than 0. Got %d",
-			params.Evidence.MaxAgeNumBlocks)
-	}
-
-	if params.Evidence.MaxAgeDuration <= 0 {
-		return errors.Errorf("evidenceParams.MaxAgeDuration must be grater than 0 if provided, Got %v",
-			params.Evidence.MaxAgeDuration)
-	}
-
-	if len(params.Validator.PubKeyTypes) == 0 {
-		return errors.New("len(Validator.PubKeyTypes) must be greater than 0")
-	}
-
-	// Check if keyType is a known ABCIPubKeyType
-	for i := 0; i < len(params.Validator.PubKeyTypes); i++ {
-		keyType := params.Validator.PubKeyTypes[i]
-		if _, ok := ABCIPubKeyTypesToAminoNames[keyType]; !ok {
-			return errors.Errorf("params.Validator.PubKeyTypes[%d], %s, is an unknown pubkey type",
-				i, keyType)
-		}
-	}
-
-	return nil
-}
-func (genDoc *GenesisDoc) SaveAs(file string) error {
-	genDocBytes, err := types.GetCodec().MarshalJSONIndent(genDoc, "", "  ")
+func getSha256Hash(file string) []byte {
+	f, err := os.Open(file)
 	if err != nil {
-		return err
+		log.Fatal(err)
 	}
-	return ioutil.WriteFile(file, genDocBytes, 0644)
+	defer f.Close()
+
+	h := sha256.New()
+	if _, err := io.Copy(h, f); err != nil {
+		log.Fatal(err)
+	}
+
+	return h.Sum(nil)
 }
