+++ cmd/export/readme.md
@@ -1,8 +1,8 @@
 # Export Command
 
 Command to convert the genesis file from version 1.0 to 1.1
-More details will be received later.
+More details will be disclosed later.
 #####Example
 ```bash
-./export --height=4350342 --chain_id=1 --genesis_time=1580206824s
+./export --height=4350342 --chain_id="minter-mainnnet-2" --genesis_time=1580206824s
 ```
